WIP added menu to structureditem
