+++ packages/frontend-2/lib/projects/graphql/fragments.ts
@@ -40,5 +40,6 @@ export const projectPageLatestItemsModelItemFragment = graphql(`
     updatedAt
     ...ProjectPageModelsCardRenameDialog
     ...ProjectPageModelsCardDeleteDialog
+    ...ProjectPageModelsActions
   }
 `)
